Fixed tests for reviews.
